# [ROCm] enable foreach fastpath