@jeffdaily 2022-03-18
> Awesome! Can `cat` workarounds in #46097 now be reverted also? The root cause for perf hit was the same (a lot of non-primitive arguments passed to the kernel)

Yes, a PR already exists for that: https://github.com/pytorch/pytorch/pull/74129